+++ cdlclient/widgets/__init__.py
@@ -12,3 +12,4 @@ This module requires `guidata`_ and `qtpy`_ packages.
 
 # pylint: disable=unused-import
 from cdlclient.widgets.connection import ConnectionDialog  # noqa: F401
+from cdlclient.widgets.objectdialog import GetObjectDialog  # noqa: F401