+++ cdlclient/widgets/svg_icons.py
@@ -1,0 +1,300 @@
+# -*- coding: utf-8 -*-
+
+"""Module containing SVG icons for the application"""
+
+SIGNAL = """<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<svg
+   role="img"
+   focusable="false"
+   aria-hidden="true"
+   viewBox="0 0 14 14"
+   version="1.1"
+   id="svg12"
+   sodipodi:docname="signal.svg"
+   inkscape:version="1.2.2 (732a01da63, 2022-12-09)"
+   xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape"
+   xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
+   xmlns="http://www.w3.org/2000/svg"
+   xmlns:svg="http://www.w3.org/2000/svg">
+  <defs
+     id="defs16" />
+  <sodipodi:namedview
+     id="namedview14"
+     pagecolor="#ffffff"
+     bordercolor="#000000"
+     borderopacity="0.25"
+     inkscape:showpageshadow="2"
+     inkscape:pageopacity="0.0"
+     inkscape:pagecheckerboard="0"
+     inkscape:deskcolor="#d1d1d1"
+     showgrid="false"
+     inkscape:zoom="42.274884"
+     inkscape:cx="3.2052128"
+     inkscape:cy="7.3093045"
+     inkscape:window-width="1920"
+     inkscape:window-height="1017"
+     inkscape:window-x="-8"
+     inkscape:window-y="-8"
+     inkscape:window-maximized="1"
+     inkscape:current-layer="svg12" />
+  <path
+     style="fill:#37474f;fill-opacity:1;stroke-width:0.0499186"
+     d="M 11.675765,12.128782 C 9.0267931,12.168326 3.6663496,12.147831 3.5042119,12.09754 2.8719874,11.901432 2.1296024,11.159709 1.872636,10.46742 1.7581758,10.159057 1.7637517,9.8882153 1.961894,6.1304014 2.1520354,2.5243182 2.1863862,2.1095818 2.3043493,1.9956358 2.4233551,1.8810227 2.8112319,1.8676583 6.4486683,1.8547498 c 3.7790137,-0.01341 4.0320977,-0.0061 4.3370687,0.1253306 0.681469,0.2936835 1.382427,1.0679658 1.550284,1.7124248 0.0337,0.129404 -0.03084,1.9300465 -0.150169,4.1891652 -0.175211,3.3172896 -0.22825,3.9853226 -0.325837,4.1041536 -0.06411,0.07807 -0.146911,0.142503 -0.184203,0.142984 z"
+     id="path468" />
+  <rect
+     style="fill:#e6e6e6;fill-opacity:1;stroke:none;stroke-width:1.57074;stroke-opacity:1"
+     id="rect10749-5"
+     width="7.8104901"
+     height="7.8187361"
+     x="3.3771245"
+     y="3.1234901"
+     transform="matrix(1,0,-0.04725952,0.99888264,0,0)" />
+  <rect
+     style="fill:#696969;fill-opacity:0.817241;stroke-width:0.402"
+     id="rect28089"
+     width="0.47309414"
+     height="7.6168156"
+     x="6.7652464"
+     y="3.1933856" />
+  <rect
+     style="fill:#696969;fill-opacity:0.817241;stroke-width:0.402"
+     id="rect28089-7"
+     width="0.47309414"
+     height="7.6168156"
+     x="6.6469727"
+     y="-10.810202"
+     transform="rotate(90)" />
+  <path
+     style="fill:none;fill-opacity:0.817241;stroke-width:0.402;stroke:#ff0000;stroke-opacity:1"
+     d="m 3.5245514,10.005941 c 0,0.09462 1.5139012,-0.5677129 1.5139012,-0.5677129 L 5.9846409,8.2081834 7.0727574,5.653475 8.2554928,4.4234302 10.195179,3.8557173"
+     id="path32265" />
+  <circle
+     cx="8.2457867"
+     cy="4.4486394"
+     r="0.58283854"
+     id="circle20-2"
+     style="fill:#ff5555;stroke-width:0.19428" />
+  <circle
+     cx="7.0560775"
+     cy="5.641458"
+     r="0.58283854"
+     id="circle22-3"
+     style="fill:#ff5555;stroke-width:0.19428" />
+  <circle
+     cx="6.0082951"
+     cy="8.2554922"
+     r="0.58283854"
+     id="circle24-6"
+     style="fill:#ff5555;stroke-width:0.19428" />
+  <circle
+     cx="5.0078225"
+     cy="9.4161282"
+     r="0.58283854"
+     id="circle26-0"
+     style="fill:#ff5555;stroke-width:0.19428" />
+  <circle
+     cx="3.534256"
+     cy="10.017958"
+     r="0.58283854"
+     id="circle28-1"
+     style="fill:#ff5555;stroke-width:0.19428" />
+  <circle
+     cx="10.147869"
+     cy="3.9030263"
+     r="0.58283854"
+     id="circle28-1-2"
+     style="fill:#ff5555;stroke-width:0.19428" />
+</svg>"""
+
+IMAGE = """<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<svg
+   role="img"
+   focusable="false"
+   aria-hidden="true"
+   viewBox="0 0 14 14"
+   version="1.1"
+   id="svg12"
+   sodipodi:docname="image.svg"
+   inkscape:version="1.2.2 (732a01da63, 2022-12-09)"
+   xml:space="preserve"
+   xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape"
+   xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
+   xmlns="http://www.w3.org/2000/svg"
+   xmlns:svg="http://www.w3.org/2000/svg"><defs
+     id="defs16" /><sodipodi:namedview
+     id="namedview14"
+     pagecolor="#ffffff"
+     bordercolor="#000000"
+     borderopacity="0.25"
+     inkscape:showpageshadow="2"
+     inkscape:pageopacity="0.0"
+     inkscape:pagecheckerboard="0"
+     inkscape:deskcolor="#d1d1d1"
+     showgrid="false"
+     inkscape:zoom="42.274884"
+     inkscape:cx="0.082791475"
+     inkscape:cy="6.0082956"
+     inkscape:window-width="1920"
+     inkscape:window-height="1017"
+     inkscape:window-x="-8"
+     inkscape:window-y="-8"
+     inkscape:window-maximized="1"
+     inkscape:current-layer="svg12" /><path
+     style="fill:#37474f;fill-opacity:1;stroke-width:0.0499186"
+     d="M 11.603869,12.127256 C 8.9548968,12.1668 3.594453,12.146305 3.4323153,12.096014 2.8000908,11.899906 2.0577058,11.158183 1.8007394,10.465894 1.6862792,10.157531 1.6918551,9.8866897 1.8899974,6.1288758 2.0801388,2.5227926 2.1144896,2.1080562 2.2324527,1.9941102 2.3514585,1.8794971 2.7393353,1.8661327 6.3767717,1.8532242 c 3.7790143,-0.01341 4.0320983,-0.0061 4.3370693,0.1253306 0.681468,0.2936835 1.382427,1.0679658 1.550284,1.7124248 0.0337,0.129404 -0.03084,1.9300465 -0.150169,4.1891652 -0.175211,3.3172892 -0.22825,3.9853222 -0.325838,4.1041532 -0.06411,0.07807 -0.14691,0.142503 -0.184203,0.142984 z"
+     id="path468" /><g
+     id="g22811"
+     transform="matrix(0.97634529,0,-0.04619032,0.97634529,0.79214309,0.58190579)"><rect
+       style="fill:#0060ff;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749"
+       width="2"
+       height="2"
+       x="2.5999999"
+       y="2.5999999" /><rect
+       style="fill:#ff7200;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7"
+       width="2"
+       height="2"
+       x="2.5999999"
+       y="4.5999999" /><rect
+       style="fill:#ff4300;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-1"
+       width="2"
+       height="2"
+       x="4.5999999"
+       y="2.5999999" /><rect
+       style="fill:#ffbb00;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7-5"
+       width="2"
+       height="2"
+       x="4.5999999"
+       y="4.5999999" /><rect
+       style="fill:#00acff;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-8"
+       width="2"
+       height="2"
+       x="6.5999999"
+       y="2.5999999" /><rect
+       style="fill:#ff3200;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7-0"
+       width="2"
+       height="2"
+       x="6.5999999"
+       y="4.5999999" /><rect
+       style="fill:#0000fb;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-1-4"
+       width="2"
+       height="2"
+       x="8.6000004"
+       y="2.5999999" /><rect
+       style="fill:#28ffcf;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7-5-5"
+       width="2"
+       height="2"
+       x="8.6000004"
+       y="4.5999999" /><rect
+       style="fill:#1fffd8;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-84"
+       width="2"
+       height="2"
+       x="2.5999999"
+       y="6.5999999" /><rect
+       style="fill:#b80000;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7-8"
+       width="2"
+       height="2"
+       x="2.5999999"
+       y="8.6000004" /><rect
+       style="fill:#0000cf;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-1-1"
+       width="2"
+       height="2"
+       x="4.5999999"
+       y="6.5999999" /><rect
+       style="fill:#40ffb7;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7-5-4"
+       width="2"
+       height="2"
+       x="4.5999999"
+       y="8.6000004" /><rect
+       style="fill:#c4ff33;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-8-8"
+       width="2"
+       height="2"
+       x="6.5999999"
+       y="6.5999999" /><rect
+       style="fill:#910000;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7-0-4"
+       width="2"
+       height="2"
+       x="6.5999999"
+       y="8.6000004" /><rect
+       style="fill:#0000ff;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-1-4-3"
+       width="2"
+       height="2"
+       x="8.6000004"
+       y="6.5999999" /><rect
+       style="fill:#44ffb3;fill-opacity:1;stroke:none;stroke-width:0.402;stroke-opacity:1"
+       id="rect10749-7-5-5-3"
+       width="2"
+       height="2"
+       x="8.6000004"
+       y="8.6000004" /></g></svg>"""
+
+GROUP = """<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<svg
+   role="img"
+   focusable="false"
+   aria-hidden="true"
+   viewBox="0 0 14 14"
+   version="1.1"
+   id="svg12"
+   sodipodi:docname="group.svg"
+   inkscape:version="1.2.2 (732a01da63, 2022-12-09)"
+   xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape"
+   xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
+   xmlns="http://www.w3.org/2000/svg"
+   xmlns:svg="http://www.w3.org/2000/svg">
+  <defs
+     id="defs16" />
+  <sodipodi:namedview
+     id="namedview14"
+     pagecolor="#ffffff"
+     bordercolor="#000000"
+     borderopacity="0.25"
+     inkscape:showpageshadow="2"
+     inkscape:pageopacity="0.0"
+     inkscape:pagecheckerboard="0"
+     inkscape:deskcolor="#d1d1d1"
+     showgrid="false"
+     inkscape:zoom="42.274884"
+     inkscape:cx="6.2330153"
+     inkscape:cy="7.2619951"
+     inkscape:window-width="1920"
+     inkscape:window-height="1017"
+     inkscape:window-x="-8"
+     inkscape:window-y="-8"
+     inkscape:window-maximized="1"
+     inkscape:current-layer="svg12" />
+  <path
+     style="fill:#37474f;fill-opacity:1;stroke-width:0.0283859"
+     d="M 6.8296648,6.2443952 C 5.1630385,5.5656538 1.8498204,4.1646697 1.7789376,4.1087109 1.5025462,3.8905091 1.4800142,3.4944286 1.7294168,3.2382259 1.8405063,3.1241066 2.0039025,3.0514944 4.345882,2.0754702 6.5932984,1.1388551 6.8595177,1.0343842 6.9999828,1.0339318 c 0.1415059,-3.633e-4 0.390015,0.09695 2.6541007,1.0402832 2.3522085,0.9800497 2.5048905,1.0479296 2.6164655,1.1632569 0.249319,0.2577033 0.226926,0.6519036 -0.04951,0.871831 -0.05551,0.044159 -1.158863,0.5197575 -2.5669417,1.1064682 C 7.5864819,6.0772998 7.159093,6.2461719 7.0283823,6.253264 6.9425113,6.257924 6.8530974,6.2539906 6.8296786,6.244414 Z"
+     id="path468" />
+  <path
+     style="fill:#e6e6e6;fill-opacity:1;stroke-width:0.0283859"
+     d="M 8.6321836,4.3315814 C 9.5064757,3.9681026 10.230488,3.6632333 10.241101,3.6540937 10.251674,3.6449595 9.5273918,3.332242 8.6315001,2.959184 L 7.0026097,2.2808951 5.3619985,2.965448 C 4.4596615,3.3419539 3.7277741,3.6507492 3.73558,3.6516593 c 0.00783,9.084e-4 0.7422977,0.3067549 1.6322008,0.6796531 0.8899054,0.3728999 1.6307817,0.6742049 1.6463938,0.6695669 0.015607,-0.00461 0.7437167,-0.3058214 1.618009,-0.6692979 z"
+     id="path470" />
+  <path
+     style="fill:#37474f;fill-opacity:1;stroke-width:0.0283859"
+     d="M 6.8296648,8.555918 C 6.7047945,8.5138635 1.9668561,6.5342623 1.8572542,6.4783472 1.7992771,6.4487701 1.72075,6.3829396 1.682745,6.3320546 1.3828407,5.9304968 1.650263,5.366658 2.1420241,5.3636968 2.2889687,5.3627885 2.4934964,5.4425038 4.6384238,6.3363819 6.122015,6.9546554 6.9974976,7.3052811 7.0370482,7.2970226 7.0713075,7.2898568 8.1340078,6.8517267 9.3986083,6.3233765 11.510477,5.4410359 11.710928,5.3628176 11.85794,5.3637005 c 0.420368,0.0025 0.706337,0.4392481 0.53752,0.8208844 C 12.31933,6.3566469 12.267463,6.4146095 12.123673,6.488199 12.061243,6.520158 10.92436,6.998606 9.5973109,7.5514122 7.1367723,8.5763955 7.0182294,8.6194195 6.8296648,8.555918 Z"
+     id="path466" />
+  <path
+     style="fill:#475b63;fill-opacity:1;stroke-width:0.0283859"
+     d="M 6.8012757,10.738642 C 6.2414245,10.520938 1.8794269,8.6854904 1.8017182,8.6349215 1.6736432,8.551578 1.5497412,8.3147523 1.5501286,8.1540355 1.5504917,8.0267197 1.653419,7.7820185 1.7384714,7.7065142 1.8523252,7.6054423 2.0631461,7.5345575 2.200317,7.551231 2.278995,7.560811 3.2500251,7.9502309 4.6555214,8.5359024 6.171646,9.1676718 7.001233,9.4994853 7.0399531,9.4896143 7.0726131,9.4812143 8.1276204,9.0454211 9.3844144,8.521022 10.755609,7.9488883 11.721549,7.5611371 11.799648,7.5514906 c 0.137204,-0.01695 0.347723,0.053715 0.461846,0.1550236 0.08505,0.075507 0.188036,0.3202055 0.188343,0.4475213 3.64e-4,0.1605988 -0.123509,0.3974956 -0.251591,0.4810624 -0.05663,0.036946 -1.213433,0.5299779 -2.5706822,1.0956284 -1.7603477,0.7336367 -2.5048743,1.0310367 -2.5973281,1.0374947 -0.071286,0.0046 -0.1743128,-0.0085 -0.2289567,-0.02957 z"
+     id="path464" />
+  <path
+     style="fill:#576e79;fill-opacity:1;stroke-width:0.0283859"
+     d="M 6.8722457,12.954603 C 6.6969628,12.909251 1.8442374,10.871475 1.7585568,10.807248 1.5332822,10.638388 1.4799938,10.238454 1.6505101,9.996368 1.6966271,9.930898 1.7918411,9.846607 1.8620975,9.809066 2.1201858,9.6711498 2.0703906,9.6546621 4.6666067,10.737631 l 2.336137,0.974492 2.3333759,-0.974308 C 11.929153,9.6551011 11.880146,9.6713468 12.137865,9.809066 c 0.07022,0.0375 0.165472,0.121825 0.211588,0.187302 0.170493,0.242052 0.11739,0.64091 -0.108041,0.81153 -0.11818,0.08944 -5.0284225,2.133784 -5.1723701,2.153474 -0.061397,0.0086 -0.1499568,0.0049 -0.196792,-0.0071 z"
+     id="path438" />
+</svg>"""